UefiPayloadPkg: Remove PEI phase from Payload
It is not necessary to have a PEI phase in the UEFI payload since no specific
PEI task is required. Instead this patch adds a UefiPayloadEntry driver to get
UEFI Payload required information from the bootloaders, converts them into a
HOB list, and loads DXE core and and transfer control to it.

Here is the change details:
1) Removed PEI phase, including Peicore, BlSupportPei, SecCore, etc.
2) Added UefiPayloadEntry driver. this is the only driver to before DXE core.
3) Added Pure X64 build support, and dropped Pure 32bit build support.
   64bit payload with 32bit entry point is still supported.
4) Use one DSC file UefiPayloadPkg.dsc to support X64 and IA32X64 build.
   Removed UefiPayloadIa32.dsc and UefiPayloadIa32X64.dsc

Tested with latest SBL and coreboot on QEMU.

Signed-off-by: Guo Dong <guo.dong@intel.com>